fix: correctly vendor charts
Previously, we allowed charts to live in a subfolder due to a missing trailing slash in the rsync call. This fixes it and moves the vendored charts to the correct path. A future sync action should be a no-op.

Note that I also removed the old ack sync method in favor of standardizing on the global sync action/workflow.