refactor: respect the upstream version and allow partial publishing with failures
We now track all vendored charts and just publish them as is to the desired chart museum.

Additionally, if any chart fails, we allow that failure to occur and continue on with publishing, exiting non-zero at the end. This allows us to publish charts that have dependencies on each other and need a dependency to be published first (ack-chart is an example).