feat: configure enhanced monitoring by setting the monitoring interval
The default value is 0, disabling enhanced monitoring. Enabling allows for collection of extra performance metrics in cloudwatch. It's slightly more expensive but necessary for anyone running a production cluster at scale..

We can allow for non-enhanced nodes in the future via helm value, but I think enabling this by default is desirable for users.