Allow setting the prometheus scaler ignoreNullValues option
The KEDA prometheus scaler has an option `ignoreNullValues` that
defaults to true. In this configuration if prometheus returns an empty
value then KEDA ignores it and treats it as a 0. This pretty much
prevents any scaling from happening.

By setting `ignoreNullValues` to false, KEDA will treat the empty
response as an error and scale to the fallback.replicas amount.

https://keda.sh/docs/2.12/scalers/prometheus/